"profile photo edit and categories edit added option added in edit of worker"

+++ Backend/src/utils/cloudinary.js
@@ -16,11 +16,16 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: "auto"
         })
         // file has been uploaded successfull
-        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+          }
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+          }
+        // remove the locally saved temporary file as the upload operation got failed
         return null;
     }
 }
